pyproject.toml
[tool.poetry]
name = "maya-stubs-distrib"
version = "0.1.0"
description = ""
authors = ["Loïc Pinsard <muream@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.2"
maya-stubgen = { path = "maya-stubgen/", develop = true }
docspec-to-jinja = { path = "docspec-to-jinja/", develop = true }
# maya-stubs = { path = "maya-stubs/", develop = true }
# maya-docs = { path = "maya-docs/", develop = true }

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pylint = "^2.14.5"

[tool.poetry.scripts]
generate-stubs = 'maya_stubgen:cli'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
typeCheckingMode = "strict"
